requirements.txt
numpy
matplotlib
scipy
astropy
astroquery
pandas
Pillow
sv-ttk
opencv-python
